move readPipedInput to input.go
