+++ cmd/ask.go
@@ -1,7 +1,6 @@
 package cmd
 
 import (
-	"bufio"
 	"context"
 	"errors"
 	"fmt"
@@ -130,33 +129,6 @@ func initializeLLMClient(logger *log.Logger) (llm.LLMClient, error) {
 	}
 
 	return llmClient, nil
-}
-
-// readPipedInput reads all input from the provided reader until EOF.
-// It's used to capture piped input from stdin.
-func readPipedInput(input io.Reader) (string, error) {
-	reader := bufio.NewReader(input)
-
-	var lines []string
-
-	for {
-		line, err := reader.ReadString('\n')
-		if err != nil {
-			if err == io.EOF {
-				if line != "" {
-					lines = append(lines, line)
-				}
-
-				break
-			}
-
-			return "", fmt.Errorf("%w: %w", ErrIO, err)
-		}
-
-		lines = append(lines, line)
-	}
-
-	return strings.Join(lines, ""), nil
 }
 
 // runSingleQuery sends a single query to the LLM and writes the response to the output.
