+++ go.mod
@@ -9,7 +9,6 @@ require github.com/google/go-cmp v0.7.0
 require (
 	github.com/charmbracelet/fang v0.4.3
 	github.com/charmbracelet/log v0.4.2
-	github.com/davecgh/go-spew v1.1.1
 	github.com/spf13/cobra v1.10.1
 	github.com/spf13/viper v1.21.0
 )
